refactor(config): move config clone into a private command module
commands/config.py was at its grandfathered loc-check ceiling (2007 code
lines); the clone command pushed it 2 lines over. Moved the command body
into a new commands/_config_clone_cmd.py, mounted via register(), mirroring
the existing commands/_storage_snapshots.py pattern. Net effect: config.py
drops the ~230-line clone command down to a 2-line import+register hook.

Ran `make loc-baseline` to record the new (lower) code-line counts for
every grandfathered file, per its own documented purpose ("after a
split"). This is that split.

WIP checkpoint -- pausing for the weekend. Remaining before this is
ready for review:
- Verify Devin/kbagent-pr-reviewer haven't already run on stale state
- Open the PR against main (branch: feat/config-clone-587)
- Request review from zajca per the user's explicit instruction
